move extension loading to tool project, improve extension.dib loading
