+++ src/Microsoft.DotNet.Interactive/DisplayExtensions.cs
@@ -14,13 +14,17 @@ public static class DisplayExtensions
     /// <param name="value">The value to display.</param>
     /// <param name="mimeTypes">The MIME types.</param>
     /// <returns>An instance of <see cref="DisplayedValue"/> that can be used to later update the display.</returns>
-    public static DisplayedValue Display(this object value,
+    public static DisplayedValue Display(
+        this object value,
         params string[] mimeTypes)
     {
         return KernelInvocationContext.Current.Display(value, mimeTypes);
     }
 
-    public static DisplayedValue DisplayAs(this string value, string mimeType, params string[] additionalMimeTypes)
+    public static DisplayedValue DisplayAs(
+        this string value, 
+        string mimeType, 
+        params string[] additionalMimeTypes)
     {
         return KernelInvocationContext.Current.DisplayAs(value, mimeType, additionalMimeTypes);
     }
